rename ipfsNode file to ipfs-node
