+++ components/browser/Post.js
@@ -1,10 +1,7 @@
-import {useState, useContext} from 'react'
-import {GlobalContext} from '../../contexts/GlobalContext'
 import {useRouter} from 'next/router'
 
 const Post = ({title, text, likes, postId}) => {
-  const [newComment, setNewComment] = useState('')
-  const {LSP7Contract, account} = useContext(GlobalContext)
+
   const router = useRouter()
 
   return(
