--- pages/index.js
@@ -1,4 +1,0 @@
-
-export default function Home({ UpAddress }) {
-  return <div className="App"></div>;
-};
