+++ pages/login.js
@@ -14,22 +14,31 @@ function Login() {
       alert('Please install UP extension or MetaMask');
       return
     }
-    // request access to the extension
-    await window.ethereum
-      .request({
-        method: 'eth_requestAccounts',
-      })
 
-      .then(function (accounts) {
-        // check if any number of accounts was returned
-        // IF go to the dashboard
-        if (accounts.length) {
-          router.push('/browse');
-          setAccount(accounts[0])
-        } else {
-          console.log('User denied access');
-        }
-      });
+    try {
+      // request access to the extension
+      await window.ethereum
+        .request({
+          method: 'eth_requestAccounts',
+        })
+
+        .then(function (accounts) {
+          // check if any number of accounts was returned
+          // IF go to the dashboard
+          if (accounts.length) {
+            router.push('/browse');
+            setAccount(accounts[0])
+          } else {
+            console.log('User denied access');
+          }
+        });
+    } catch (error) {
+      if(error.message === 'User denied access'){
+        console.log('User denied access');
+      } else {
+        console.log(error);
+      }
+    }
   }
 
   useEffect(() => {
